test(signals): split detectorsPhase tests into sibling file
Move the 4 detectorsPhase test cases from
tests/lib/orchestration/post-merge-pipeline.test.js into a new sibling
tests/lib/orchestration/detectors-phase.test.js, mirroring the
production split (detectorsPhase now lives in
.agents/scripts/lib/orchestration/detectors-phase.js).

Keeps the parent test file focused on phase orchestration assertions
and recovers post-merge-pipeline.test.js's maintainability score
within tolerance of its prior baseline.

baseline-refresh: ratchet maintainability for the test split.
