+++ baselines/maintainability.json
@@ -517,6 +517,7 @@
   "tests/lib/orchestration/concurrent-task-resolver.test.js": 88.289,
   "tests/lib/orchestration/context.test.js": 121.009,
   "tests/lib/orchestration/crap-remediation-1641.test.js": 107.184,
+  "tests/lib/orchestration/detectors-phase.test.js": 108.832,
   "tests/lib/orchestration/dispatch-pipeline-runWorktreeGc.test.js": 109.936,
   "tests/lib/orchestration/epic-cleanup.test.js": 99.815,
   "tests/lib/orchestration/epic-lifecycle-detector.test.js": 105.011,
@@ -527,7 +528,7 @@
   "tests/lib/orchestration/phase-runner.test.js": 121.235,
   "tests/lib/orchestration/plan-runner/worktree-sweep.test.js": 94.608,
   "tests/lib/orchestration/post-merge-pipeline-stale-registry.test.js": 109.766,
-  "tests/lib/orchestration/post-merge-pipeline.test.js": 103.321,
+  "tests/lib/orchestration/post-merge-pipeline.test.js": 104.417,
   "tests/lib/orchestration/retro-append-checks.test.js": 102.808,
   "tests/lib/orchestration/retro-runner.test.js": 91.294,
   "tests/lib/orchestration/story-close/baseline-attribution-wiring.test.js": 100.604,
